# smaller `Task` object