With this `sizeof(jl_task_t)` goes from 616 bytes to ~~416~~ 352. Who likes free performance?

master, `pfib(31)`, 4 threads:
```
  1.879936 seconds (13.24 M allocations: 1.902 GiB, 41.12% gc time)
```
This PR:
```
  1.617563 seconds (13.22 M allocations: 1.499 GiB, 36.26% gc time)
```

master, spawn/wait test from #36699, 4 threads:
```
3.090069 seconds (21.87 M allocations: 2.872 GiB, 2.63% gc time)
```
this PR:
```
2.852884 seconds (21.90 M allocations: 2.218 GiB, 3.54% gc time)
```

I would consider it unlikely/inadvisable for a task to hold 29 locks :joy: 
I'm planning to remove a couple more redundant fields from `Task`, but this should be a big obvious win for now.